use cast instead of compiler option in fi
git-svn-id: file:///home/amuse/svn/amuse/trunk@3664 6e832d89-e7d1-4ad1-8ca1-873beecccf65

+++ src/amuse/community/fi/src/fcco.f90
@@ -110,7 +110,7 @@ end function
 function fccoco(n,G0,Z,Pe)
  real :: n,G0,Z,Pe,xia,fccoco
 
- xia=e2solve(4./3.*n/1.d6/G0*(1+24.8*Z)) 
+ xia=e2solve(REAL(4./3.*n/1.d6/G0*(1+24.8*Z))) 
  fccoco=MAX(0.d0,1-xiA*1.d3/n0/xifuv/1.66/Z/sqrt(Pe/1.d4))**3
 end function
